Separated slide up and slide down unit tests to separate file
