+++ CMakeLists.txt
@@ -527,6 +527,7 @@ set(HWY_TEST_FILES
   hwy/tests/reduction_test.cc
   hwy/tests/reverse_test.cc
   hwy/tests/shift_test.cc
+  hwy/tests/slide_up_down_test.cc
   hwy/tests/swizzle_test.cc
   hwy/tests/test_util_test.cc
   hwy/tests/tuple_test.cc
